--- resources/ts/abstract/Adder.ts
@@ -1,3 +1,0 @@
-export default abstract class {
-    public abstract injectContent(): void;
-}
