--- resources/ts/abstract/Modifier.ts
@@ -1,3 +1,0 @@
-export default abstract class {
-    public abstract modify(): void;
-}
