+++ resources/ts/classes/Adders/HomeLinksAdder.ts
@@ -3,7 +3,7 @@ import conf from '@/conf'
 import AdderInterface from '@/classes/Adders/Adder'
 
 export default class implements AdderInterface {
-    public injectContent(): void {
+    public add(): void {
         const target = document.querySelector(conf.selectors.targetForHomeLinks)
 
         if (window.location.pathname !== '/' || !target)
