+++ resources/ts/classes/Adders/SearchIconAdder.ts
@@ -2,7 +2,7 @@ import conf from '@/conf'
 import AdderInterface from '@/classes/Adders/Adder'
 
 export default class implements AdderInterface {
-    public injectContent(): void {
+    public add(): void {
         const target = document.querySelector(conf.selectors.targetForSearchIcon)
 
         if (!target)
