+++ resources/ts/classes/Extension.ts
@@ -1,53 +1,22 @@
-import conf from '@/conf'
-import CodeSampleModifier from '@/classes/CodeSampleModifier'
-import Modifier from '@/abstract/Modifier'
-import RandomVideoAdder from '@/classes/Adders/RandomVideoAdder'
-import Adder from "@/abstract/Adder"
-import HomeLinksAdder from '@/classes/Adders/HomeLinksAdder'
-import SearchIconAdder from '@/classes/Adders/SearchIconAdder'
-import LogoModifier from '@/classes/Modifiers/LogoModifier'
+import type Adder from '@/classes/Adders/Adder'
+import type Modifier from '@/classes/Modifiers/Modifier'
 
 export default class {
-    public static focusOnTheSearchBarOnHomePage(): void {
-        const searchBar = document.querySelector<HTMLInputElement>(conf.selectors.searchQuery)
-
-        if (searchBar && window.location.pathname === '/')
-            searchBar.focus()
+    /**
+     * Adders are classes that add (insert) elements to the page.
+     * Things like links, buttons, etc.
+     */
+    public applyAdders(adders: Adder[]): this {
+        adders.forEach(adder => adder.add())
+        return this
     }
 
-    public static applyFunctionPageModifiers(): void {
-        const codeAreas = document.querySelectorAll<HTMLDivElement>(conf.selectors.codeExamples)
-        const staticClasses = document.querySelectorAll<HTMLDivElement>(conf.selectors.classMethods)
-        const phpVersion = document.querySelector<HTMLParagraphElement>(conf.selectors.phpVersionInfo)
-
-        if (!codeAreas || !staticClasses || !phpVersion)
-            return
-
-        const modifiers: Array<Modifier> = [
-            new CodeSampleModifier(codeAreas, staticClasses),
-        ]
-
-        for (const modifier of modifiers)
-            modifier.modify()
-    }
-
-    public static applyHomePageAdders(): void {
-        const adders: Adder[] = [
-            new RandomVideoAdder(),
-            new HomeLinksAdder(),
-            new SearchIconAdder(),
-        ]
-
-        for (const adder of adders)
-            adder.injectContent()
-    }
-
-    public static applyModifiers(): void {
-        const modifiers: Modifier[] = [
-            new LogoModifier(),
-        ]
-
-        for (const modifier of modifiers)
-            modifier.modify()
+    /**
+     * Modifiers are classes that modify elements on the page.
+     * Things like changing the logo, fixing styles, etc.
+     */
+    public applyModifiers(modifiers: Modifier[]): this {
+        modifiers.forEach(modifier => modifier.modify())
+        return this
     }
 }
